requirements/base.txt
# 基础依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0

# 数据库和存储
nebula3-python>=3.5.0
qdrant-client>=1.7.0

# AI和机器学习 - 专用模型版本
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
tokenizers>=0.14.0
wandb>=0.15.0

# 深度学习框架
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 模型训练支持
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.15.0

# 工具库
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# 日志和监控
structlog>=23.2.0

# 异步支持
aiofiles>=23.2.1
asyncio-mqtt>=0.16.1